update smooth to return double

+++ core/src/main/java/com/airbnb/aerosolve/core/function/Function.java
@@ -28,9 +28,9 @@ public interface Function extends Serializable, Cloneable {
 
   // for function not support smooth, just return 0
   // for function support smooth, return errAndCoeff
-  float smooth(double tolerance);
+  double smooth(double tolerance);
   // for function support smooth, return errAndCoeff percentage
-  float smoothByTolerancePercentage(double tolerancePercentage);
+  double smoothByTolerancePercentage(double tolerancePercentage);
 
   Function clone() throws CloneNotSupportedException;
 }
